\name{NRF1pair.bam}
\alias{NRF1pair.bam}
\alias{NRF1pair.bai}
\alias{EnrichmentRegions_mm10.rds}
\alias{ReferenceMethylation.rds}
\alias{AllCs.rds}

\title{SingleMoleculeFootprintingData}

\description{
This Data package contains r objcets necessary to run some of the functions from the SingleMoleculeFootprinting package.
SingleMoleculeFootprinting is an R package providing functions to analyze Single Molecule Footprinting (SMF) data.
}

\usage{
NRF1pair.bam(metadata = FALSE)
NRF1pair.bai(metadata = FALSE)
EnrichmentRegions_mm10.rds(metadata = FALSE)
ReferenceMethylation.rds(metadata = FALSE)
AllCs.rds(metadata = FALSE)
}

\arguments{
  \item{metadata}{FALSE (default) returns data. TRUE returns metadata}
}
